embeddings/pyproject.toml
[project]
name = "aithena-embeddings"
version = "0.1.0"
description = "Aithena Embedding Services"
readme = "README.md"
authors = [
    { name = "Camilo Velez", email = "camilov61@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "grpcio-tools>=1.73.1",
    "numpy>=2.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src/polus"]

[tool.hatch.build.targets.wheel]
include = ["src/polus"]

[tool.hatch.build.targets.wheel.sources]
"src/polus" = "polus"
